test(capi): consolidate integration tests
