+++ crates/capi/src/lib.rs
@@ -14,6 +14,9 @@ use std::{
     ptr,
 };
 
+#[cfg(test)]
+use cc as _;
+
 mod alloc;
 
 type CStyleReadFileCallback = unsafe extern "C" fn(
