+++ crates/capi/tests/it/ctest.rs
@@ -1,5 +1,3 @@
-#![allow(unused_crate_dependencies)]
-
 use std::{
     env,
     ffi::OsString,
@@ -15,11 +13,6 @@ fn c_api_smoke_test() {
         return;
     }
 
-    let Some(cc) = c_compiler() else {
-        eprintln!("skipping C API smoke test because no C compiler was found");
-        return;
-    };
-
     let manifest_dir = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
     let workspace_dir = manifest_dir.join("../..");
     let target_dir = env::var_os("CARGO_TARGET_DIR")
@@ -29,13 +22,16 @@ fn c_api_smoke_test() {
 
     let mut build = Command::new(env::var_os("CARGO").unwrap_or_else(|| OsString::from("cargo")));
     build.current_dir(&workspace_dir).args(["build", "-p", "solar-capi", "--lib"]);
-    assert_command(build, "build solar-capi cdylib");
+    crate::assert_command(build, "build solar-capi cdylib");
 
     let out_dir = target_dir.join("ctest");
     fs::create_dir_all(&out_dir).unwrap();
     let exe = out_dir.join("solar-capi-ctest");
 
-    let mut compile = Command::new(cc);
+    let Some(mut compile) = c_compiler() else {
+        eprintln!("skipping C API smoke test because no C compiler was found");
+        return;
+    };
     compile
         .arg("-I")
         .arg(manifest_dir.join("include"))
@@ -46,22 +42,31 @@ fn c_api_smoke_test() {
         .arg(format!("-Wl,-rpath,{}", lib_dir.display()))
         .arg("-o")
         .arg(&exe);
-    assert_command(compile, "compile C API smoke test");
+    crate::assert_command(compile, "compile C API smoke test");
 
     let mut run = Command::new(&exe);
     prepend_dynamic_library_path(&mut run, &lib_dir);
-    assert_command(run, "run C API smoke test");
+    crate::assert_command(run, "run C API smoke test");
 }
 
-fn c_compiler() -> Option<OsString> {
-    if let Some(cc) = env::var_os("CC") {
-        return Some(cc);
+fn c_compiler() -> Option<Command> {
+    let target = target_triple()?;
+    let mut build = cc::Build::new();
+    build.cargo_metadata(false).warnings(false).target(&target).host(&target).opt_level(0);
+    build.try_get_compiler().ok().map(|tool| tool.to_command())
+}
+
+fn target_triple() -> Option<String> {
+    if let Ok(target) = env::var("TARGET") {
+        return Some(target);
     }
-    ["cc", "clang", "gcc"].into_iter().find(|cc| command_exists(cc)).map(OsString::from)
-}
 
-fn command_exists(command: &str) -> bool {
-    Command::new(command).arg("--version").output().is_ok()
+    let output = Command::new(env::var_os("RUSTC").unwrap_or_else(|| OsString::from("rustc")))
+        .arg("-vV")
+        .output()
+        .ok()?;
+    let stdout = String::from_utf8(output.stdout).ok()?;
+    stdout.lines().find_map(|line| line.strip_prefix("host: ")).map(str::to_owned)
 }
 
 fn prepend_dynamic_library_path(command: &mut Command, lib_dir: &Path) {
@@ -71,16 +76,4 @@ fn prepend_dynamic_library_path(command: &mut Command, lib_dir: &Path) {
         paths.extend(env::split_paths(&existing));
     }
     command.env(key, env::join_paths(paths).unwrap());
-}
-
-fn assert_command(mut command: Command, label: &str) {
-    let output = command.output().unwrap_or_else(|error| panic!("failed to {label}: {error}"));
-    if !output.status.success() {
-        panic!(
-            "failed to {label}\nstatus: {}\nstdout:\n{}\nstderr:\n{}",
-            output.status,
-            String::from_utf8_lossy(&output.stdout),
-            String::from_utf8_lossy(&output.stderr)
-        );
-    }
 }
